requirements.txt
pandas
Pillow
python-dotenv
openai
streamlit